% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savvy_glm2.R
\name{savvy_glm2}
\alias{savvy_glm2}
\title{Generalized Linear Models with Slab and Shrinkage Estimators}
\usage{
savvy_glm2(formula, family = gaussian, data, weights,
               model_class = c("St", "DSh", "SR", "GSR"), subset,
               na.action, start = NULL, etastart, mustart, offset,
               control = list(...), model = TRUE,
               method = "savvy_glm.fit2", x = FALSE, y = TRUE,
               contrasts = NULL, use_parallel = TRUE, ...)
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the model to be fitted. As for \code{\link{glm2}}.}

\item{family}{A description of the error distribution and link function to be used in the model. As for \code{\link{glm2}}.}

\item{data}{An optional data frame, list or environment containing the variables in the model. As for \code{\link{glm2}}.}

\item{weights}{An optional vector of weights to be used in the fitting process. As for \code{\link{glm2}}.}

\item{model_class}{A character vector specifying the shrinkage method(s) to be used in the underlying fitter \code{savvy_glm.fit2}.
Allowed values are \code{"St"}, \code{"DSh"}, \code{"SR"}, \code{"GSR"}, and \code{"Sh"}. If a single value is provided, only that method is executed.
If multiple values are provided, the specified methods are evaluated (in parallel if \code{use_parallel = TRUE}) and the one with the lowest AIC is returned.
By default, the value is \code{c("St", "DSh", "SR", "GSR")}; note that the \code{"Sh"} method is considered only if explicitly included.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process. As for \code{\link{glm2}}.}

\item{na.action}{A function which indicates what should happen when the data contain NAs. As for \code{\link{glm2}}.}

\item{start}{Starting values for the parameters in the linear predictor. As for \code{\link{glm2}}.}

\item{etastart}{Starting values for the linear predictor. As for \code{\link{glm2}}.}

\item{mustart}{Starting values for the vector of means. As for \code{\link{glm2}}.}

\item{offset}{An optional vector specifying a priori known component to be included in the linear predictor during fitting. As for \code{\link{glm2}}.}

\item{control}{A list of control parameters to pass to the iterative fitting process. As for \code{\link{glm2}}.}

\item{model}{A logical value indicating whether the model frame should be included as a component of the returned value. As for \code{\link{glm2}}.}

\item{method}{The method to be used in fitting the model. The default is \code{savvy_glm.fit2}. This method uses IRLS
with custom optimization methods to ensure better convergence by evaluating different fitting methods and selecting the best one based on AIC.
As in \code{\link{glm2}}, the alternative method "model.frame" returns the model frame and does no fitting.}

\item{x}{A logical value indicating whether the model matrix used in the fitting process should be returned as a component of the returned value. As for \code{\link{glm2}}.}

\item{y}{A logical value indicating whether the response vector used in the fitting process should be returned as a component of the returned value. As for \code{\link{glm2}}.}

\item{contrasts}{An optional list. See the contrasts.arg of \code{model.matrix.default}. As for \code{\link{glm2}}.}

\item{use_parallel}{A logical value specifying whether to evaluate multiple shrinkage methods in parallel.
Defaults to \code{TRUE}. Setting this to \code{FALSE} forces sequential evaluation.}

\item{...}{Additional arguments to be passed to the low level regression fitting functions. As for \code{\link{glm2}}.}
}
\value{
The value returned by \code{savvy_glm2} has exactly the same structure as that returned by \code{glm2}, except for:
\item{method}{the name of the fitter function used, which by default is \code{savvy_glm.fit2}.}
\item{chosen_fit}{the name of the chosen fitting method based on AIC.}
}
\description{
\code{savvy_glm2} extends the classical \code{glm2} function from the \code{glm2} package by embedding a set of shrinkage-based methods within the iteratively reweighted least squares (IRLS) algorithm.
These shrinkage methods (implemented via \code{savvy_glm.fit2}) are designed to improve convergence and estimation accuracy.
The user can specify one or more methods through the \code{model_class} argument. When multiple methods are provided (default is \code{c("St", "DSh", "SR", "GSR")}),
the function can evaluate them in parallel (controlled by the \code{use_parallel} argument) and selects the final model based on the lowest Akaike Information Criterion (AIC).
}
\details{
\code{savvy_glm2} improves upon the standard Generalized Linear Model (GLM) fitting process by incorporating shrinkage estimator functions
(\code{St_ost}, \code{DSh_ost}, \code{SR_ost}, \code{GSR_ost}, and optionally \code{Sh_ost}) within the IRLS algorithm.
The function begins with initial parameter estimates and updates the coefficients using the specified shrinkage methods.
When multiple methods are specified in \code{model_class}, the function may evaluate them in parallel (when \code{use_parallel = TRUE})
and then selects the final model based on the lowest AIC, ensuring that the model converges to the best possible solution given the data and specified family.
}
\references{
Marschner, I. C. (2011). \emph{glm2: Fitting Generalized Linear Models with Convergence Problems}.
The R Journal, 3(2), 12–15. doi:10.32614/RJ-2011-012. Available at:
\url{https://doi.org/10.32614/RJ-2011-012}.

Asimit, V., Cidota, M. A., Chen, Z., & Asimit, J. (2025). \emph{Slab and Shrinkage Linear Regression Estimation}.
Retrieved from \url{https://openaccess.city.ac.uk/id/eprint/35005/}.

SavvyGLM article. \url{https://ziwei-chenchen.github.io/savvyGLM}.
}
\seealso{
\code{\link{glm}}, \code{\link{glm2}}, \code{\link{savvy_glm.fit2}}
}
\author{
Ziwei Chen and Vali Asimit\cr
Maintainer: Ziwei Chen <ziwei.chen.3@citystgeorges.ac.uk>
}
